requirements.txt
# Core dependencies
# Note: pydub code is included in pydub_plus/core (forked from original pydub)
numpy>=1.21.0

# GPU acceleration (optional)
cupy-cuda12x>=12.0.0; sys_platform != "darwin"  # CUDA 12.x (adjust for your CUDA version)
# cupy-cuda11x>=11.0.0  # For CUDA 11.x
# cupy-cuda10x>=10.0.0  # For CUDA 10.x

# Async processing
aiofiles>=23.0.0
asyncio-throttle>=1.0.2

# REST API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# CLI tools
click>=8.1.7
rich>=13.7.0
typer>=0.9.0

# YouTube/TikTok workflows
yt-dlp>=2023.11.16
moviepy>=1.0.3

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0

